restructured component files
